Fix some typos in the README and improve error handling

+++ TNTGitHook/misc/tnt_git_hook.sh
@@ -35,7 +35,11 @@ then
 
   REMOTE=$(git ls-remote --get-url | head -1)
 
-  python3 -m TNTGitHook --commit-msgs-file $FILENAME --remote $REMOTE
+  if ! python3 -m TNTGitHook --commit-msgs-file $FILENAME --remote $REMOTE;
+  then
+    echo "Error executing python hook"
+    exit 1
+  fi
   rm $FILENAME
 fi
 cd $CURRENT_PATH || exit 0
